# src/BoxServerApi/Controllers/BoxController.cs
/*
 * BoxServer API
 *
 * API and objects for BoxServer
 *
 * OpenAPI spec version: 1.0.0
 *
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using System.ComponentModel.DataAnnotations;
using BoxServer.Models;
using BoxServer.Interfaces;

namespace BoxServer.Controllers
{
    /// <summary>
    ///
    /// </summary>
    [ApiController]
    public class BoxController : ControllerBase
    {
        private readonly IBoxRepository _boxRepository;
        private readonly ILogger<BoxController> _logger;

        /// <summary>
        /// constructor
        /// </summary>
        /// <param name="logger"></param>
        /// <param name="boxRepository"></param>
        public BoxController(ILogger<BoxController> logger, IBoxRepository boxRepository)
        {
            _logger = logger;
            _boxRepository = boxRepository;
        }

        /// <summary>
        /// Delete an existing box
        /// </summary>
        /// <param name="clientId"></param>
        /// <param name="id">Id</param>
        /// <response code="400">Invalid ID supplied</response>
        /// <response code="404">Box not found</response>
        /// <response code="405">Validation exception</response>
        [HttpDelete]
        [Route("/api/v1/{clientId}/box/{id:guid}")]
        [SwaggerOperation("DeleteBox")]
        public async Task<IActionResult> DeleteBox([FromRoute][Required] string clientId, [FromRoute][Required] Guid id)
        {
            var result = await _boxRepository.DeleteBox(clientId, id).ConfigureAwait(false);
            return Ok(result);
        }

        /// <summary>
        /// Get a list of boxs
        /// </summary>
        /// <response code="200">successful operation</response>
        [HttpGet]
        [Route("/api/v1/{clientId}/box")]
        [SwaggerOperation("GetBoxs")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<Box>), description: "successful operation")]
        public async Task<ActionResult<List<Box>>> GetBox([FromRoute] [Required] string clientId)
        {
            var allBoxs = (await _boxRepository.GetBoxs(clientId).ConfigureAwait(false))?.ToList();
            if (!allBoxs?.Any() ?? false)
            {
                return NotFound();
            }
            return Ok(allBoxs);
        }

        /// <summary>
        /// Get a box by Id
        /// </summary>
        /// <param name="id">Id</param>
        /// <param name="clientId"></param>
        /// <response code="200">successful operation</response>
        /// <response code="400">Invalid ID supplied</response>
        /// <response code="405">Validation exception</response>
        [HttpGet]
        [Route("/api/v1/{clientId}/box/{id:guid}")]
        [SwaggerOperation("GetBox")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<Box>), description: "successful operation")]
        public async Task<ActionResult<Box>> GetBox([FromRoute][Required] Guid id, [FromRoute][Required] string clientId)
        {
            var ret = await _boxRepository.GetBox(clientId, id).ConfigureAwait(false);
            if (ret == null)
            {
                return NotFound();
            }
            return Ok(ret);
        }

        /// <summary>
        /// Create a new box
        /// </summary>
        /// <param name="box">Box object that needs to be added</param>
        /// <param name="clientId"></param>
        /// <response code="400">Invalid ID supplied</response>
        /// <response code="405">Validation exception</response>
        [HttpPost]
        [Route("/api/v1/{clientId}/box")]
        [SwaggerOperation("NewBox")]
        public async Task<ActionResult<Box>> NewBox([FromBody] Box box, [FromRoute][Required] string clientId)
        {
            if (box.BoxId is not null && box.BoxId != Guid.Empty)
            {
                return BadRequest("Must not have id on new");
            }

            return await _boxRepository.AddBox(clientId, box).ConfigureAwait(false);
        }

        /// <summary>
        /// Update an existing box to create a new version
        /// </summary>
        /// <param name="box">Box object that needs to be updated</param>
        /// <param name="clientId"></param>
        /// <response code="400">Invalid ID supplied</response>
        /// <response code="404">Box not found</response>
        /// <response code="405">Validation exception</response>
        [HttpPut]
        [Route("/api/v1/{clientId}/box")]
        [SwaggerOperation("UpdateBox")]
        public async Task<ActionResult<Box?>> UpdateBox([FromBody] Box box, [FromRoute][Required] string clientId)
        {
            if (box.BoxId == Guid.Empty)
            {
                return BadRequest("Must have id on update");
            }
            return await _boxRepository.UpdateBox(clientId, box).ConfigureAwait(false);
        }
    }
}